broke up benchmarks more
