+++ rangeset.cabal
@@ -105,6 +105,16 @@ benchmark complexity-bench
                        containers,
   main-is:             ComplexityBench.hs
 
+benchmark contiguity-bench
+  import:              benchmark-common
+  type:                exitcode-stdio-1.0
+  build-depends:       containers,
+                       QuickCheck,
+                       array,
+                       selective
+  main-is:             ContiguityBench.hs
+  --other-modules:       Data.EnumSet
+
 source-repository head
   type:                git
   location:            https://github.com/j-mie6/rangeset
